Explicit cast to int
After introduction of float values to options the conversion of unsigned int to int or vice versa is regarded as the same rank as conversion to and from float. Fix this by using an explicit cast

+++ src/apps/common/stream_options.cpp
@@ -86,8 +86,8 @@ int StreamKeyValueParser::updateConfiguration(CameraConfiguration *config,
 		StreamConfiguration &cfg = config->at(i++);
 
 		if (opts.isSet("width") && opts.isSet("height")) {
-			cfg.size.width = opts["width"];
-			cfg.size.height = opts["height"];
+			cfg.size.width = static_cast<int>(opts["width"]);
+			cfg.size.height = static_cast<int>(opts["height"]);
 		}
 
 		if (opts.isSet("pixelformat"))
